requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
python-dotenv==1.0.1
Werkzeug==3.0.3
requests==2.31.0
numpy==1.26.4
